+++ src/dataset/infrastructure/repositories/DatasetJSDataverseRepository.ts
@@ -45,6 +45,7 @@ import {
   updateTermsOfAccess,
   updateDatasetLicense,
   getDatasetUploadLimits,
+  getDatasetReviews,
   DatasetType,
   getDatasetAvailableDatasetTypes
 } from '@iqss/dataverse-client-javascript'
@@ -66,6 +67,7 @@ import { AxiosResponse } from 'axios'
 import { JSDataverseReadErrorHandler } from '@/shared/helpers/JSDataverseReadErrorHandler'
 import { CollectionSummary } from '@/collection/domain/models/CollectionSummary'
 import { DatasetUploadLimits } from '@/dataset/domain/models/DatasetUploadLimits'
+import { DatasetReview } from '@/dataset/domain/models/DatasetReview'
 
 const includeDeaccessioned = true
 
@@ -484,5 +486,9 @@ export class DatasetJSDataverseRepository implements DatasetRepository {
 
   getDatasetUploadLimits(datasetId: string | number): Promise<DatasetUploadLimits> {
     return getDatasetUploadLimits.execute(datasetId)
+  }
+
+  getDatasetReviews(datasetId: string | number): Promise<DatasetReview[]> {
+    return getDatasetReviews.execute(datasetId)
   }
 }