+++ src/sections/dataset/Dataset.tsx
@@ -32,6 +32,7 @@ import { DatasetPublishingStatus } from '@/dataset/domain/models/Dataset'
 import { DataverseInfoRepository } from '@/info/domain/repositories/DataverseInfoRepository'
 import { useAnonymized } from './anonymized/AnonymizedContext'
 import { useDatasetRepositories } from '@/shared/contexts/repositories/RepositoriesProvider'
+import { DatasetReviews } from './dataset-reviews/DatasetReviews'
 
 interface DatasetProps {
   fileRepository: FileRepository
@@ -148,6 +149,7 @@ export function Dataset({
               {(!isCurrentVersionDeaccessioned || canUpdateDataset) && (
                 <DatasetMetrics data-testid="dataset-metrics" datasetId={dataset.persistentId} />
               )}
+              <DatasetReviews datasetId={dataset.persistentId} />
             </Col>
           </Row>